core API for reboost
