+++ src/reboost/__init__.py
@@ -3,4 +3,4 @@ from __future__ import annotations
 from reboost import optmap
 from reboost._version import version as __version__
 
-__all__ = ["__version__", "build_hit", "core", "optmap"]
+__all__ = ["__version__", "build_hit", "core", "iterator", "optmap"]
